Platemap toolbar

See merge request research/atlas!14

+++ ui/src/components/PlateMapToolbar.js
@@ -1,42 +1,84 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Form, Checkbox, Button } from 'semantic-ui-react';
+import { connect } from 'react-redux';
+import { Button, Popup } from 'semantic-ui-react';
 
+import {
+  createExperimentActions,
+  addNewPlateMap,
+  clonePlateMap,
+  selectActivePlateMap,
+} from '../store/createExperiment';
 import { PlateMapDropdown } from './PlateMapDropdown';
 import styles from './PlateMapToolbar.module.css';
+import { ClonePlateForm } from './ClonePlateForm';
 
-export class PlateMapToolbar extends Component {
-  handleDelete = () => {
-    this.props.onDeleteClick(this.props.activePlateMap.id);
+class PlateMapToolbar extends Component {
+  state = {
+    clonePopupOpen: false,
+  };
+  handleAddClick = () => {
+    if (this.props.onAddClick) {
+      this.props.onAddClick();
+    }
+  };
+  handleDeleteClick = () => {
+    if (this.props.onDeleteClick) {
+      this.props.onDeleteClick(this.props.activePlateMap.id);
+    }
   };
   handleHighlight = (e, { value }) => {
-    this.props.onHighlightClick({ componentType: value });
+    if (this.props.onHighlightClick) {
+      this.props.onHighlightClick({ componentType: value });
+    }
+  };
+  handleCloneSubmit = ({ typesToClone }) => {
+    if (this.props.onCloneSubmit) {
+      this.props.onCloneSubmit(this.props.activePlateMap.id, typesToClone);
+    }
+    this.handleClonePopupClose();
+  };
+  handleClonePopupOpen = () => {
+    this.setState({ clonePopupOpen: true });
+  };
+  handleClonePopupClose = () => {
+    this.setState({ clonePopupOpen: false });
   };
   render() {
-    const { plateMaps, activePlateMap, highlightedComponents } = this.props;
+    const { plateMaps, activePlateMap } = this.props;
     return (
       <div className={styles.toolbar}>
-        <div>
+        <div className={styles.dropdownContainer}>
           <PlateMapDropdown
             plateMaps={plateMaps}
             activePlateMap={activePlateMap}
             onChange={this.props.onPlateMapChange}
           />
         </div>
-        <div>
+        <div className={styles.buttonsContainer}>
           <Button
             primary
             icon="plus circle"
-            content="Add Plate Map"
-            onClick={this.props.onAddClick}
+            content="Add Plate"
+            onClick={this.handleAddClick}
           />
           <Button
             icon="trash"
-            content="Delete Plate Map"
-            onClick={this.handleDelete}
+            content="Delete Plate"
+            onClick={this.handleDeleteClick}
           />
+          <Popup
+            trigger={<Button icon="clone" content="Clone Plate" />}
+            on="click"
+            open={this.state.clonePopupOpen}
+            position="bottom left"
+            onClose={this.handleClonePopupClose}
+            onOpen={this.handleClonePopupOpen}
+          >
+            <ClonePlateForm onSubmit={this.handleCloneSubmit} />
+          </Popup>
         </div>
-        <div className={styles.highlight}>
+        {/* <div className={styles.highlight}>
           <Form>
             <Form.Group inline>
               <label>Highlight wells containing:</label>
@@ -66,7 +108,7 @@ export class PlateMapToolbar extends Component {
               </Form.Field>
             </Form.Group>
           </Form>
-        </div>
+        </div> */}
       </div>
     );
   }
@@ -76,7 +118,29 @@ PlateMapToolbar.propTypes = {
   plateMaps: PropTypes.array.isRequired,
   activePlateMap: PropTypes.object.isRequired,
   highlightedComponents: PropTypes.array.isRequired,
-  onDeleteClick: PropTypes.func.isRequired,
-  onHighlightClick: PropTypes.func.isRequired,
-  onAddClick: PropTypes.func.isRequired,
+  onPlateMapChange: PropTypes.func,
+  onDeleteClick: PropTypes.func,
+  onHighlightClick: PropTypes.func,
+  onAddClick: PropTypes.func,
+  onCloneSubmit: PropTypes.func,
 };
+
+const mapState = (state, props) => {
+  const { plateMaps, highlightedComponents } = state.createExperiment;
+  const activePlateMap = selectActivePlateMap(state);
+  return { plateMaps, highlightedComponents, activePlateMap };
+};
+
+const mapDispatch = {
+  onPlateMapChange: createExperimentActions.setActivePlateMap,
+  onDeleteClick: createExperimentActions.deletePlateMap,
+  onHighlightClick: createExperimentActions.toggleHighlight,
+  onAddClick: addNewPlateMap,
+  onCloneSubmit: clonePlateMap,
+};
+
+const connected = connect(
+  mapState,
+  mapDispatch
+)(PlateMapToolbar);
+export { connected as PlateMapToolbar };
